add tabIndex, remove textenumeration

+++ src/Shared/Core/Controls/TabStrip/EventArgs/TabStripItemTappedEventArgs.cs
@@ -2,11 +2,11 @@
 {
     public class TabStripItemTappedEventArgs
     {
-        public TabStripItemTappedEventArgs(int index)
+        public TabStripItemTappedEventArgs(byte index)
         {
             Index = index;
         }
 
-        public int Index { get; }
+        public byte Index { get; }
     }
 }
